# Assignment_3/docs/rentals-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Rental Search API",
    "description": "This API provides access to publicly available data on the Australian rental market in 2026. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    }
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": ["Authentication"],
        "description": "Creates a new user account.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "examples": ["mike@gmail.com"] },
                  "password": { "type": "string", "examples": ["password"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["User created"] }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["Request body incomplete - email and password needed"] }
                  }
                }
              }
            }
          },
          "500": {
            "description": "User already exists or server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "examples": [false] },
                    "message": { "type": "string", "examples": ["User already exists"] }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["Authentication"],
        "description": "Log in to an existing user account.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "examples": ["mike@gmail.com"] },
                  "password": { "type": "string", "examples": ["password"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string", "examples": ["ajsonwebtoken"] },
                    "tokenType": { "type": "string", "examples": ["Bearer"] },
                    "expiresIn": { "type": "number", "examples": [86400] }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["Request body incomplete - email and password needed"] }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["User does not exist"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Passwords do not match"] }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/debugLogin": {
      "post": {
        "tags": ["Authentication"],
        "description": "Log in with a 1-second expiry token for testing purposes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "examples": ["mike@gmail.com"] },
                  "password": { "type": "string", "examples": ["password"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful with short expiry token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string", "examples": ["ajsonwebtoken"] },
                    "tokenType": { "type": "string", "examples": ["Bearer"] },
                    "expiresIn": { "type": "number", "examples": [1] }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["Request body incomplete - email and password needed"] }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["User does not exist"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Passwords do not match"] }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/{email}/profile": {
      "get": {
        "tags": ["User"],
        "description": "Returns the profile of the specified user. Authenticated users viewing their own profile will also receive dob and address fields.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "description": "Unauthenticated or non-matching user response",
                      "type": "object",
                      "properties": {
                        "email": { "type": "string" },
                        "firstName": { "type": ["string", "null"] },
                        "lastName": { "type": ["string", "null"] }
                      },
                      "required": ["email", "firstName", "lastName"]
                    },
                    
                    {
                      "description": "Authenticated matching user response",
                      "type": "object",
                      "properties": {
                        "email": { "type": "string" },
                        "firstName": { "type": ["string", "null"] },
                        "lastName": { "type": ["string", "null"] },
                        "dob": { "type": ["string", "null"], "examples": ["1990-01-01"] },
                        "address": { "type": ["string", "null"] }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["User not found"] }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "description": "Updates the profile of the specified user. Requires authorisation and the authenticated user must match the profile being updated.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "dob": { "type": "string", "examples": ["1990-01-01"] },
                  "address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "dob": { "type": "string", "examples": ["1990-01-01"] },
                    "address": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Request body incomplete: firstName, lastName, dob and address are required."] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Request body invalid: firstName, lastName and address must be strings only."] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Invalid input: dob must be a real date in format YYYY-MM-DD."] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Invalid input: dob must be a date in the past."] }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Authorization header ('Bearer token') not found"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["JWT token has expired"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Invalid JWT token"] }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - cannot update another user's profile.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["Forbidden: you can only update your own profile"] }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}